boulder/cli: Implement update command to update version in recipes
Updates an existing recipe (stone.yml) with a new version and tarball,
provided by the user. Additionally, it recomputes the hash of the new
tarball and increments the release.

PoC only because dyaml is a shitter and doesn't provide any formatting
options when dumping a yaml file to disk making the thing look my left ball
and my big toe had a baby.